refactor: move `Media.jsx` contents into `media.jsx`
